remove the Update class, replace its usage by StateMachine::Update
The idea of the Update class in the elmo driver was to get
more "aggregate" views on what has been updated. Since the
object-to-bitfield was done internally to the driver itself,
let them handle it and only keep the object-level API in
canopen_master.

Add a template-based API to Update for a more consistent manipulation
of Update in a library that uses Slave and the object definitions
as types.